[FEAT] 아이디어 생성 요청 API

+++ src/main/java/channeling/be/domain/report/domain/Report.java
@@ -53,10 +53,10 @@ public class Report extends BaseEntity {
     private Integer concept; //컨셉 일관성
 
     @Column
-    private Long seo; // seo 구성
+    private Integer seo; // seo 구성
 
     @Column
-    private Long revisit; // 재방문률
+    private Integer revisit; // 재방문률
 
     @Column(columnDefinition = "TEXT")
     private String summary; // 요약본
